Conflicts:
	db/data/accounting.mv.db
	src/main/java/ir/visoft/accounting/ui/controller/BillManagementController.java
	src/main/java/ir/visoft/accounting/ui/controller/EditBillController.java
	src/main/java/ir/visoft/accounting/util/PdfUtil.java
	src/main/resources/config.properties

+++ src/patch/patch1.sql
@@ -35,7 +35,7 @@ CREATE TABLE BILL (
   previous_figure INT,
   current_figure INT,
   cunsumption INT,
-  abonman DOUBLE ,
+  abonman INT ,
   reduction INT,
   services INT,
   cost_water DOUBLE,
